+++ src/QueryCat.Backend.Core/Functions/FunctionDelegate.cs
@@ -1,9 +1,0 @@
-using QueryCat.Backend.Core.Execution;
-using QueryCat.Backend.Core.Types;
-
-namespace QueryCat.Backend.Core.Functions;
-
-/// <summary>
-/// QueryCat function signature.
-/// </summary>
-public delegate VariantValue FunctionDelegate(IExecutionThread thread);