+++ src/QueryCat.Samples/Collection/CustomInlineFunctionUsage.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using QueryCat.Backend;
+using QueryCat.Backend.Core.Execution;
 using QueryCat.Backend.Core.Types;
 
 namespace QueryCat.Samples.Collection;
@@ -10,7 +11,7 @@ internal class CustomInlineFunctionUsage : BaseUsage
     public override async Task RunAsync()
     {
         var executionThread = new ExecutionThreadBootstrapper().Create();
-        executionThread.FunctionsManager.RegisterFunction("secret(a: string, b: numeric): string", thread =>
+        executionThread.FunctionsManager.RegisterFunction("secret(a: string, b: numeric): string", (IExecutionThread thread) =>
         {
             var a = thread.Stack[0];
             var b = thread.Stack[1];