+++ src/QueryCat.UnitTests/Functions/DefaultFunctionsManagerTests.cs
@@ -21,7 +21,7 @@ public sealed class DefaultFunctionsManagerTests
     }
 
     [Fact]
-    public void RegisterFunction_SimpleFunction_ShouldEvalCorrectly()
+    public async Task RegisterFunction_SimpleFunction_ShouldEvalCorrectly()
     {
         // Act.
         _functionsManager.RegisterFunction(SumIntegers);
@@ -29,7 +29,7 @@ public sealed class DefaultFunctionsManagerTests
         var func = _functionsManager.FindByName("add",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.Integer, DataType.Integer));
         using var frame = Executor.Thread.Stack.CreateFrameWithArguments(new VariantValue(5), new VariantValue(6));
-        var ret = func.Delegate(Executor.Thread);
+        var ret = await FunctionCaller.CallAsync(func.Delegate, Executor.Thread);
 
         Assert.Equal(11L, ret.AsIntegerUnsafe);
     }
@@ -41,7 +41,7 @@ public sealed class DefaultFunctionsManagerTests
     }
 
     [Fact]
-    public void RegisterFunction_VariadicArgumentMethod_ShouldEvalCorrectly()
+    public async Task RegisterFunction_VariadicArgumentMethod_ShouldEvalCorrectly()
     {
         // Act.
         _functionsManager.RegisterFunction(SumIntegersVariadic);
@@ -51,13 +51,13 @@ public sealed class DefaultFunctionsManagerTests
                 DataType.String, DataType.Integer, DataType.Integer));
         using var frame = Executor.Thread.Stack.CreateFrameWithArguments(
             new VariantValue("sum"), new VariantValue(5), new VariantValue(5));
-        var ret = func.Delegate(Executor.Thread);
+        var ret = await FunctionCaller.CallAsync(func.Delegate, Executor.Thread);
 
         Assert.Equal("sum: 10", ret);
     }
 
     [Fact]
-    public void RegisterFunction_VariadicArgumentMethodEmpty_ShouldEvalCorrectly()
+    public async Task RegisterFunction_VariadicArgumentMethodEmpty_ShouldEvalCorrectly()
     {
         // Act.
         _functionsManager.RegisterFunction(SumIntegersVariadic);
@@ -65,7 +65,7 @@ public sealed class DefaultFunctionsManagerTests
         var func = _functionsManager.FindByName("add",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.String));
         using var frame = Executor.Thread.Stack.CreateFrameWithArguments(new VariantValue("sum"));
-        var ret = func.Delegate(Executor.Thread);
+        var ret = await FunctionCaller.CallAsync(func.Delegate, Executor.Thread);
 
         Assert.Equal("sum: 0", ret);
     }
@@ -83,7 +83,7 @@ public sealed class DefaultFunctionsManagerTests
     }
 
     [Fact]
-    public void RegisterFunction_RegisterFromClass_ShouldEvalCorrectly()
+    public async Task RegisterFunction_RegisterFromClass_ShouldEvalCorrectly()
     {
         // Arrange.
         _functionsManager.RegisterFromType(typeof(TestClass1));
@@ -91,13 +91,13 @@ public sealed class DefaultFunctionsManagerTests
         // Act.
         var func1 = _functionsManager.FindByName("test_class1",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.Integer, DataType.String));
-        var value1 = Executor.Thread.CallFunction(func1.Delegate, 1, "2");
+        var value1 = await FunctionCaller.CallWithArgumentsAsync(func1.Delegate, Executor.Thread, [1, "2"]);
         var func2 = _functionsManager.FindByName("test_class2",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.Integer, DataType.String));
-        var value2 = Executor.Thread.CallFunction(func2.Delegate, 1, "2");
+        var value2 = await FunctionCaller.CallWithArgumentsAsync(func2.Delegate, Executor.Thread, [1, "2"]);
         var func3 = _functionsManager.FindByName("test_class3",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.Integer, DataType.String));
-        var value3 = Executor.Thread.CallFunction(func3.Delegate, 1, "2");
+        var value3 = await FunctionCaller.CallWithArgumentsAsync(func3.Delegate, Executor.Thread, [1, "2"]);
 
         // Assert.
         Assert.Equal("1 2", value1.As<TestClass1>()!.Value);
@@ -119,17 +119,17 @@ public sealed class DefaultFunctionsManagerTests
     }
 
     [Fact]
-    public void RegisterFunction_RegisterFromType_ShouldEvalCorrectly()
+    public async Task RegisterFunction_RegisterFromType_ShouldEvalCorrectly()
     {
         // Arrange.
         _functionsManager.RegisterFromType(typeof(TestClass2));
 
         // Act.
         var func1 = _functionsManager.FindByName("function1");
-        Executor.Thread.CallFunction(func1.Delegate);
+        await FunctionCaller.CallWithArgumentsAsync(func1.Delegate, Executor.Thread);
         var func2 = _functionsManager.FindByName("testfunc",
             FunctionCallArgumentsTypes.FromPositionArguments(DataType.String));
-        var value2 = Executor.Thread.CallFunction(func2.Delegate, "2");
+        var value2 = await FunctionCaller.CallWithArgumentsAsync(func2.Delegate, Executor.Thread, ["2"]);
 
         // Assert.
         Assert.Equal(2, value2.AsInteger);